orient function uses exif_read_data for which valid image types are only JPEG and TIFF as per the document [1].
When trying to upload an screenshot with png type I was getting message as *failed* even though image was uploading. So added a fix to check image type before calling orient function. Not sure if this is a proper fix but after this i'm not getting any error during file upload

[1] http://au2.php.net/manual/en/function.exif-read-data.php