French regions spelling
* Updated spelling of French regions, using proper Unicode letters.
* Still using the pre-2016 divisions though, while waiting to see
whether the new divisions are sticking
https://fr.wikipedia.org/wiki/R%C3%A9gion_fran%C3%A7aise

Note that Guadeloupe, Martinique, Guyane, Reunion, and Mayotte are
currently registered as states, but they are official French regions.
This seems inconsistent with e.g. Alaska and Hawai which are listed
under USA.
I can fix if you want.

Note 2: The format of this file seems to be quite difficult to maintain
(e.g. to add/remove a state) due to integer indexes.