fix for probe loop in case that network of contact has changed
This fixes a loop of probes / contact updates where the network has been apup and for whatever
reason is now feed. In this case the contact hasn't been updated and rescheduled with next_update as null resulting
in a huge number of queries.

The call to hasLocalData has been moved a bit up as it checks for values that otherwise would be emptied befare the call.

Fixes #13037